+++ pyo3-stub-gen/src/stub_type/builtins.rs
@@ -1,0 +1,31 @@
+use crate::stub_type::*;
+
+macro_rules! impl_builtin {
+    ($ty:ty, $pytype:expr) => {
+        impl PyStubType for $ty {
+            fn type_output() -> TypeInfo {
+                TypeInfo {
+                    name: $pytype.to_string(),
+                    import: HashSet::new(),
+                }
+            }
+        }
+    };
+}
+
+impl_builtin!(bool, "bool");
+impl_builtin!(u8, "int");
+impl_builtin!(u16, "int");
+impl_builtin!(u32, "int");
+impl_builtin!(u64, "int");
+impl_builtin!(u128, "int");
+impl_builtin!(i8, "int");
+impl_builtin!(i16, "int");
+impl_builtin!(i32, "int");
+impl_builtin!(i64, "int");
+impl_builtin!(i128, "int");
+impl_builtin!(f32, "float");
+impl_builtin!(f64, "float");
+
+impl_builtin!(&str, "str");
+impl_builtin!(String, "str");
